chore: move some misc utils out from utils file
